+++ src/simulation/events.jl
@@ -423,30 +423,31 @@ function removeHostFromPopulation!(host_idx::Int64, population::Population, mode
     deleteat!(population.hosts, host_idx)
 end
 
-function setPopulationContactCoefficient!(pop_idx_1::Int64, pop_idx_2::Int64, coefficient::Float64, model::Model)
-    model.populations[pop_idx_1].population_contact_coefficients[pop_idx_2] = coefficient
+function setPopulationContactCoefficient!(pop_1::Population, pop_2::Population, coefficient::Float64, model::Model)
+    pop_1.population_contact_coefficients[ model.population_dict[pop_2.id] ] = coefficient
     change = (
-        model.populations[pop_idx_1].population_contact_coefficients[pop_idx_2] *
-        model.population_weights_receive[RECEIVE_CONTACT-CHOICE_MODIFIERS[1]+1, pop_idx_2] /
+        pop_1.population_contact_coefficients[ model.population_dict[pop_2.id] ] *
+        model.population_weights_receive[ RECEIVE_CONTACT-CHOICE_MODIFIERS[1]+1, model.population_dict[pop_2.id] ] /
         max(
-            length(model.populations[pop_idx_2].hosts) *
-            model.populations[pop_idx_2].parameters.constant_contact_density,
+            length(pop_2.hosts) *
+            pop_2.parameters.constant_contact_density,
             1.0
-        )) - model.population_contact_weights_receive[pop_idx_2, pop_idx_1]
-    updatePopulationContactWeightReceiveMatrix!(pop_idx_1, pop_idx_2, change, model)
+        )) - model.population_contact_weights_receive[ model.population_dict[pop_2.id] ,  model.population_dict[pop_1.id] ]
+    updatePopulationContactWeightReceiveMatrix!(model.population_dict[pop_1.id], model.population_dict[pop_2.id], change, model)
 end
 
-function setPopulationTransitionCoefficient!(pop_idx_1::Int64, pop_idx_2::Int64, coefficient::Float64, model::Model)
-    model.populations[pop_idx_1].population_transition_coefficients[pop_idx_2] = coefficient
+function setPopulationTransitionCoefficient!(pop_1::Population, pop_2::Population, coefficient::Float64, model::Model)
+    pop_1.population_transition_coefficients[ model.population_dict[pop_2.id] ] = coefficient
     change = (
-        model.populations[pop_idx_1].population_transition_coefficients[pop_idx_2] *
-        model.population_weights_receive[RECEIVE_TRANSITION-CHOICE_MODIFIERS[1]+1, pop_idx_2] /
+        pop_1.population_transition_coefficients[ model.population_dict[pop_2.id] ] *
+        model.population_weights_receive[ RECEIVE_TRANSITION-CHOICE_MODIFIERS[1]+1, model.population_dict[pop_2.id] ] /
         max(
-            length(model.populations[pop_idx_2].hosts) *
-            model.populations[pop_idx_2].parameters.constant_transition_density,
+            length(pop_2.hosts) *
+            pop_2.parameters.constant_transition_density,
             1.0
-        )) - model.population_transition_weights_receive[pop_idx_2, pop_idx_1]
-    updatePopulationTransitionWeightReceiveMatrix!(pop_idx_1, pop_idx_2, change, model)
+        )) - model.population_transition_weights_receive[ model.population_dict[pop_2.id], model.population_dict[pop_1.id] ]
+    println(change)
+    updatePopulationTransitionWeightReceiveMatrix!(model.population_dict[pop_1.id], model.population_dict[pop_2.id], change, model)
 end
 
 # Model events